requirements.txt
setuptools
pip
numpy
scipy
scikit-image
argparse
configparser
opencv-python
pandas
nose
coverage
torch
torchvision
tensorboard
terminaltables
virtualenv
reedsolo
pycodestyle
sphinx
sphinx_markdown_tables
sphinx_rtd_theme
beautifulsoup4
requests
hydra-core
typing
omegaconf
tqdm
Cython
Chamaeleo